+++ kant/events/exceptions.py
@@ -1,6 +1,0 @@
-class EventDoesNotExist(Exception):
-    pass
-
-
-class EventError(Exception):
-    pass