+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Removed
 - Remove event dependencies
+- Remove declarative events
 
 ### Changed
 - Improves the contribute guide